This merges the android14-6.1 branch into the -lts branch, catching
it up with the latest changes in there.

It contains the following commits:

* 9e207186c745 Merge tag 'android14-6.1.138_r00' into android14-6.1
* 3c6d0251e1fb ANDROID: ABI: Update pixel symbol list
* ed6999107ec6 ANDROID: vendor_hook: add trace_android_rvh_setscheduler_prio
* 5b71d364254c ANDROID: binder: fix minimum node priority comparison
* 785e57725850 BACKPORT: KVM: arm64: Eagerly switch ZCR_EL{1,2}
* 6a31e426c64e BACKPORT: KVM: arm64: Calculate cptr_el2 traps on activating traps
* 89720e9e1bc3 BACKPORT: KVM: arm64: Mark some header functions as inline
* b9b8d84f6cfc BACKPORT: KVM: arm64: Refactor exit handlers
* c00c44bea22d BACKPORT: KVM: arm64: Remove VHE host restore of CPACR_EL1.SMEN
* c952e23cf8df BACKPORT: KVM: arm64: Remove VHE host restore of CPACR_EL1.ZEN
* a08391468f2f BACKPORT: KVM: arm64: Remove host FPSIMD saving for non-protected KVM
* 12921b6e2348 BACKPORT: KVM: arm64: Unconditionally save+flush host FPSIMD/SVE/SME state
* f1df93017ee7 BACKPORT: KVM: arm64: Discard any SVE state when entering KVM guests
* 21c687a8c532 ANDROID: KVM: arm64: Eagerly restore host FPSIMD/SVE state in pKVM
* d871a6444c97 ANDROID: KVM: arm64: Move __deactivate_fpsimd_traps() to switch.h
* 1b3dfc7c3845 ANDROID: KVM: arm64: Move kvm_hyp_handle_fpsimd_host() to switch.h
* c3b505e78c57 ANDROID: KVM: arm64: Remove pkvm_set_max_sve_vq()
* d653b32842d8 Revert "ANDROID: KVM: arm64: Use enum instead of helper for fp state"
* b07be5e5113b ANDROID: GKI: update symbol list file for xiaomi
* 58b3f63bc69f ANDROID: vendor_hooks: Add hooks for pcp related optimization.
* ad7902a401f6 BACKPORT: mm: userfaultfd: correct dirty flags set for both present and swap pte
* e30317e116ff ANDROID: 16K: Remove ELF padding entry from map_file ranges
* 228e0f23bdeb UPSTREAM: net_sched: sch_sfq: move the limit validation
* 3e7cb920f167 UPSTREAM: net_sched: sch_sfq: use a temporary work area for validating configuration
* a0fa2316cce1 ANDROID: ABI: Update pixel symbol list
* 218e2bd24587 FROMGIT: perf/aux: Allocate non-contiguous AUX pages by default
* 3cd01bb5bd37 UPSTREAM: mm: Fix is_zero_page() usage in try_grab_page()
* 53b26534cce7 UPSTREAM: usb-storage: Optimize scan delay more precisely
* 646380b087a5 UPSTREAM: usb: gadget: f_midi: Fixing wMaxPacketSize exceeded issue during MIDI bind retries
* 1cc1e931722a UPSTREAM: fork: avoid inappropriate uprobe access to invalid mm
* c5abfe08fc65 UPSTREAM: fork: do not invoke uffd on fork if error occurs
* 7c2011337f5b UPSTREAM: KVM: arm64: Ensure vgic_ready() is ordered against MMIO registration
* 34f1eb99850e UPSTREAM: mm/mglru: only clear kswapd_failures if reclaimable
* 98cb57aeb332 UPSTREAM: serial: qcom-geni: fix receiver enable
* a368123b90e5 UPSTREAM: serial: qcom-geni: fix dma rx cancellation
* 7192539e3e36 UPSTREAM: serial: qcom-geni: revert broken hibernation support
* b45e2c927411 UPSTREAM: PM: domains: Fix alloc/free in dev_pm_domain_attach|detach_list()
* b29cc3971e26 UPSTREAM: usb: dwc3: core: Fix compile warning on s390 gcc in dwc3_get_phy call
* 48ab183a3ee2 ANDROID: GKI: Update symbol list for bcmstb
* ca2f65da73b1 UPSTREAM: codel: remove sch->q.qlen check before qdisc_tree_reduce_backlog()
* 7b89b57429c2 ANDROID: mm/memfd-ashmem-shim: Simplify buffer name retrieval
* 0c1a07d9c284 ANDROID: GKI: Update the symbol list for mtk
* 242f90b45e89 UPSTREAM: drm/msm: Rename drm_msm_gem_submit_reloc::or in C++ code
* 81ea45b132e6 ANDROID: GKI: Update symbol list for mtk
* 1c4f8fb026ac UPSTREAM: net_sched: Prevent creation of classes with TC_H_ROOT

+++ drivers/usb/dwc3/core.c
@@ -1480,7 +1480,7 @@ static int dwc3_core_get_phy(struct dwc3 *dwc)
 	struct phy		*temp_phy = NULL;
 	char phy_name[9];
 	int ret;
-	int i;
+	u8 i;
 
 	if (node) {
 		dwc->usb2_phy = devm_usb_get_phy_by_phandle(dev, "usb-phy", 0);
@@ -1510,7 +1510,7 @@ static int dwc3_core_get_phy(struct dwc3 *dwc)
 		if (vdwc->num_usb2_ports == 1)
 			snprintf(phy_name, sizeof(phy_name), "usb2-phy");
 		else
-			snprintf(phy_name, sizeof(phy_name),  "usb2-%d", i);
+			snprintf(phy_name, sizeof(phy_name),  "usb2-%u", i);
 
 		temp_phy = devm_phy_get(dev, phy_name);
 		if (IS_ERR(temp_phy)) {
@@ -1532,7 +1532,7 @@ static int dwc3_core_get_phy(struct dwc3 *dwc)
 		if (vdwc->num_usb3_ports == 1)
 			snprintf(phy_name, sizeof(phy_name), "usb3-phy");
 		else
-			snprintf(phy_name, sizeof(phy_name), "usb3-%d", i);
+			snprintf(phy_name, sizeof(phy_name), "usb3-%u", i);
 
 		temp_phy = devm_phy_get(dev, phy_name);
 		if (IS_ERR(temp_phy)) {
